+++ app/src/main/java/io/github/sds100/keymapper/system/network/NetworkAdapter.kt
@@ -22,4 +22,10 @@ interface NetworkAdapter {
     fun disableMobileData(): Result<*>
 
     fun getKnownWifiSSIDs(): List<String>?
+
+    suspend fun sendHttpRequest(
+        method: HttpMethod,
+        url: String,
+        body: String,
+    ): Result<*>
 }